requirements.txt
aiogram
numpy
pandas
python-dotenv==1.0.0
rapidfuzz==3.5.2